☄ Switch to webview/webview
Wails refuses to work, I tried a svelte with typescript setup first, worked at the beginning then started giving me weird errors in the console from core-js (!!). Removing the typescript stuff did nothing to change things.
As a final try I tried to wails init again from scratch with the *vanilla* HTML/CSS/JS preset, which would not launch either (not even modifying anything from the default template)

So I'm switching to a simple webview, I'll have 2 things runnning:

- The webview, listening to a localhost server on some port
- The Go backend, sending responses to that localhost server

The drawback is that the backend can only send data to respond to a request, not send a request to e.g. notify of changes in the local filesystem. That may be possible with websockets though, and would be a good opportunity to finally learn them.