♿️ Make toolbar buttons non-focusable
The toolbar will have keyboard shortcuts (that should be announced with aria-labelledby as the shortcuts will be on buttons' tooltips), and removing them from focus access allows tabbing from keys to values in <FieldMap> and start typing immediately. It's much less of a pain when it comes to actually starting to type stuff out.